+++ operator.ps1
@@ -148,4 +148,5 @@ Set-Alias -name ml -value Get-OGP
 . $pwshSrcDir/i_function.ps1
 . $pwshSrcDir/tenki_function.ps1
 . $pwshSrcDir/sleepy_function.ps1
+. $pwshSrcDir/Get-AppShortcut_function.ps1