+++ operator.ps1
@@ -97,6 +97,7 @@ if ($IsWindows){
 
 
 ## writing
+. $pwshSrcDir/mdgrep_function.ps1
 . $pwshSrcDir/tex2pdf_function.ps1
 . $pwshSrcDir/jl_function.ps1
 . $pwshSrcDir/kinsoku_function.ps1