+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Updated: [README.md]
+- Updated [README.md]
+- Fixed [percentile] synopsis
 
 ### Added
 
-- Added: [Get-AppShortcut] fucntion
+- Added [mdgrep] function
+- Added [Get-AppShortcut] function
 
 
 ## [3.0.1] - 2023-03-17 Fri
@@ -233,6 +235,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [fval]: src/fval_function.ps1
 
 [Get-AppShortcut]: src/Get-AppShortcut_function.ps1
+[mdgrep]: src/mdgrep_function.ps1
 
 
 [unreleased]: https://github.com/btklab/posh-mocks/compare/3.0.1..HEAD