+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added [mdgrep] function
 - Added [Get-AppShortcut] function
-
+- Added [clip2img] examples
 
 ## [3.0.1] - 2023-03-17 Fri